1. Refactored query parameter based functions to services/users file for reusability. 2. Updated the integration tests after refactoring of code.
